+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+Added:
+
+* Allow passing an `Array` of arguments to `Parser#parse()`
+
 Changed:
 
 * Replace dependency [minimist](https://github.com/substack/minimist) with